requirements.txt
Flask==2.2.2
roboflow
opencv-python
Pillow
